modified: tipos dos arrays modificados para double

+++ Algoritmos/InsertionSort.java
@@ -2,13 +2,13 @@ package Algoritmos;
 
 public class InsertionSort {
     
-    public void Insertion_Sort(int[] vetor) {
+    public void Insertion_Sort(double[] vetor) {
 
         Integer tamanho = vetor.length;
 
         for (int i = 1; i < tamanho; i++) {
 
-            int numeroAtual = vetor[i];
+            double numeroAtual = vetor[i];
             int indexAnterior = i-1;
 
             while ((indexAnterior >= 0) && (vetor[indexAnterior] > numeroAtual)) {
